% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{add_example_mode}
\alias{add_example_mode}
\title{Add a mode to a prompt (example function)}
\usage{
add_example_mode(prompt_wrap_or_list)
}
\arguments{
\item{prompt_wrap_or_list}{A single string, a prompt_wrap object, or a list
of prompt_wrap objects.}
}
\value{
A prompt list with an added mode prompt wrapper object.
}
\description{
Add a mode to a prompt (example function)
}
