% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text.R
\name{add_text}
\alias{add_text}
\title{Add text to a tidyprompt}
\usage{
add_text(prompt, text, position = c("after", "before"), sep = "\\n\\n")
}
\arguments{
\item{prompt}{A single string or a \code{\link[=tidyprompt]{tidyprompt()}} object}

\item{text}{Text to be added to the current prompt text}

\item{position}{Where to add the text; either "after" or "before".}

\item{sep}{Separator to be used between the current prompt text and the text to be added}
}
\value{
A \code{\link[=tidyprompt]{tidyprompt()}} with an added \code{\link[=prompt_wrap]{prompt_wrap()}} which
will append the text to the end of the current prompt text
}
\description{
Add text to a prompt by adding a \code{\link[=prompt_wrap]{prompt_wrap()}} which will append the text to
the before or after the current prompt text.
}
\seealso{
Other other_prompt_wraps: 
\code{\link{set_system_prompt}()}
}
\concept{other_prompt_wraps}
