% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tools.R
\name{add_tools_add_documentation}
\alias{add_tools_add_documentation}
\title{Add tidyprompt function documentation to a function}
\usage{
add_tools_add_documentation(
  func,
  description,
  arguments = list(),
  return_value,
  name = NULL
)
}
\arguments{
\item{func}{A function object}

\item{description}{A description of the function}

\item{arguments}{A named list of arguments (arguments) with descriptions}

\item{return_value}{A description of the return value}

\item{name}{The name of the function (optional). If not provided, the function
name will be extracted from the function object}
}
\value{
The function object with the documentation added as an attribute
}
\description{
This function adds documentation to a custom function. This documentation
is used to extract information about the function's name, description, arguments,
and return value. This information is used to provide an LLM with information
about the functions, so that the LLM can call R functions. The intended
use of this function is to add documentation to custom functions that do not
have help files; \code{\link[=add_tools_get_documentation]{add_tools_get_documentation()}} will extract the documentation
from help files when available.
}
\examples{
my_function <- function(x, y) { x + y }
my_function_documented <- add_tools_add_documentation(
   my_function,
   description = "Add two numbers",
   arguments = list(
     x = "A number",
     y = "Another number"
   ),
   return_value = "The sum of the two numbers"
 )
}
