% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_as_code.R
\name{answer_as_code}
\alias{answer_as_code}
\title{Instruct LLM to answer a prompt with R code}
\usage{
answer_as_code(
  prompt,
  add_text = "You must code in the programming language 'R' to answer this prompt.",
  pkgs_to_use = c(),
  evaluate_code = TRUE,
  evaluation_session = NULL,
  list_packages = TRUE,
  list_objects = TRUE,
  skim_dataframes = TRUE,
  output_as_tool = FALSE,
  return_mode = c("full", "code", "console", "object", "formatted_output", "llm_answer")
)
}
\arguments{
\item{prompt}{A character string or a tidyprompt object}

\item{add_text}{Character string which will be added to the prompt text,
informing the LLM that they must code in R to answer the prompt.}

\item{pkgs_to_use}{A character vector of package names that may be used
in the R code that the LLM will generate. If evaluating the R code, these
will be pre-loaded in the R session.}

\item{evaluate_code}{Logical indicating whether the R code should be
evaluated. If TRUE, the R code will be evaluated in a new R session
(using the 'callr' package).}

\item{evaluation_session}{A pre-existing r_session object (from the 'callr' package)
to evaluate the R code (e.g., with certain objects loaded). If NULL, a
new r_session object will be created.}

\item{list_packages}{Logical indicating whether the LLM should be informed
about the packages that may be used in the R code (if TRUE, a list of the
loaded packages will be shown in the initial prompt)}

\item{list_objects}{Logical indicating whether the LLM should be informed
about the objects that already exist in the R session (if TRUE, a list of the objects
plus their types will be shown in the initial prompt)}

\item{skim_dataframes}{Logical indicating whether the LLM should be informed
about the structure of dataframes that already exist in the R session (if TRUE,
a skim summary of each dataframe will be shown in the initial prompt)}

\item{output_as_tool}{Logical indicating whether the console output of the
evaluated R code should be sent back to the LLM, meaning the LLM will use
R code as a tool to formulate an answer to the prompt. If TRUE, the LLM
can decide if they can answer the prompt with the output, or if they need to modify
their R code. Once the LLM does not provide new R code (i.e., the prompt is being answered)
this prompt wrap will end (it will continue for as long as the LLM provides R code).}

\item{return_mode}{Character string indicating the return mode. One of
'full', 'code', 'console', 'object', 'formatted_output', or 'llm_answer'.
If 'full', the function will return a list with the original LLM answer,
the extracted R code, and (if evaluated) the output of the R code.
If 'code', the function will return the extracted R code. If 'console', the
function will return the console output of the evaluated R code.
If 'object', the function will return the object produced by the evaluated R code.
If 'formatted_output', the function will return a formatted string with the
extracted R code, its console output, and a print of the last object (identical
to how it would be presented to the LLM if 'output_as_tool' is TRUE).
If 'llm_answer', the function will return only the original LLM answer.
When choosing 'console' or 'object', an additional instruction will be added to
the prompt text to inform the LLM about the expected output of the R code.}
}
\value{
A tidyprompt object with the new prompt wrap added to it
}
\description{
This function adds a prompt wrap to a tidyprompt object that instructs the
LLM to answer the prompt with R code. There are various options to customize
the behavior of this prompt wrap, concerning the evaluation of the R code,
the packages that may be used, the objects that already exist in the R
session, and if the console output that should be sent back to the LLM.
}
\details{
For the evaluation of the R code, the 'callr' package is required.
Please note: automatic evaluation of generated R code may be dangerous to your
system; you must use this function with caution.
}
\seealso{
\code{\link[=add_tools]{add_tools()}}

Other answer_as: 
\code{\link{answer_as_boolean}()},
\code{\link{answer_as_integer}()},
\code{\link{answer_as_list}()},
\code{\link{answer_as_named_list}()},
\code{\link{answer_as_regex}()}
}
\concept{answer_as}
