% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_as_integer.R
\name{answer_as_integer}
\alias{answer_as_integer}
\title{Make LLM answer as an integer (between min and max)}
\usage{
answer_as_integer(
  prompt,
  min = NULL,
  max = NULL,
  add_instruction_to_prompt = TRUE
)
}
\arguments{
\item{prompt}{A single string or a \code{\link[=tidyprompt]{tidyprompt()}} object}

\item{min}{(optional) Minimum value for the integer}

\item{max}{(optional) Maximum value for the integer}

\item{add_instruction_to_prompt}{(optional) Add instruction for replying
as an integer to the prompt text. Set to FALSE for debugging if extractions/validations
are working as expected (without instruction the answer should fail the
validation function, initiating a retry)}
}
\value{
A \code{\link[=tidyprompt]{tidyprompt()}} with an added \code{\link[=prompt_wrap]{prompt_wrap()}} which
will ensure that the LLM response is an integer.
}
\description{
Make LLM answer as an integer (between min and max)
}
\seealso{
Other answer_as: 
\code{\link{answer_as_boolean}()},
\code{\link{answer_as_code}()},
\code{\link{answer_as_list}()},
\code{\link{answer_as_named_list}()},
\code{\link{answer_as_regex}()}
}
\concept{answer_as}
