% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_as_list.R
\name{answer_as_list}
\alias{answer_as_list}
\title{Make LLM answer as a list of items}
\usage{
answer_as_list(prompt, item_name = "item", n_unique_items = NULL)
}
\arguments{
\item{prompt}{A single string or a \code{\link[=tidyprompt]{tidyprompt()}} object}

\item{item_name}{(optional) Name of the items in the list}

\item{n_unique_items}{(optional) Number of unique items required in the list}
}
\value{
A \code{\link[=tidyprompt]{tidyprompt()}} with an added \code{\link[=prompt_wrap]{prompt_wrap()}} which
will ensure that the LLM response is a list of items
}
\description{
Make LLM answer as a list of items
}
\seealso{
\code{\link[=answer_as_named_list]{answer_as_named_list()}}

Other answer_as: 
\code{\link{answer_as_boolean}()},
\code{\link{answer_as_code}()},
\code{\link{answer_as_integer}()},
\code{\link{answer_as_named_list}()},
\code{\link{answer_as_regex}()}
}
\concept{answer_as}
