% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_as_named_list.R
\name{answer_as_named_list}
\alias{answer_as_named_list}
\title{Extract named list from LLM response with optional item instructions and validations}
\usage{
answer_as_named_list(
  prompt,
  item_names,
  item_instructions = NULL,
  item_validations = NULL
)
}
\arguments{
\item{prompt}{A single string or a \code{\link[=tidyprompt]{tidyprompt()}} object}

\item{item_names}{A character vector specifying the expected item names}

\item{item_instructions}{An optional named list of additional instructions for each item}

\item{item_validations}{An optional named list of validation functions for each item.
Like validation functions for a \code{\link[=prompt_wrap]{prompt_wrap()}}, these functions should return
\code{\link[=llm_feedback]{llm_feedback()}} if the validation fails. If the validation
is successful, the function should return TRUE}
}
\value{
A \code{\link[=tidyprompt]{tidyprompt()}} with an added \code{\link[=prompt_wrap]{prompt_wrap()}} that ensures
the LLM response is a named list with the specified item names, optional
instructions, and validations.
}
\description{
Extract named list from LLM response with optional item instructions and validations
}
\seealso{
\code{\link[=answer_as_list]{answer_as_list()}} \code{\link[=llm_feedback]{llm_feedback()}}

Other answer_as: 
\code{\link{answer_as_boolean}()},
\code{\link{answer_as_code}()},
\code{\link{answer_as_integer}()},
\code{\link{answer_as_list}()},
\code{\link{answer_as_regex}()}
}
\concept{answer_as}
