% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_by.R
\name{answer_by_chain_of_thought}
\alias{answer_by_chain_of_thought}
\title{Set chain of thought mode for a prompt}
\usage{
answer_by_chain_of_thought(prompt, extract_from_finish_brackets = TRUE)
}
\arguments{
\item{prompt}{A tidyprompt object or a single string}

\item{extract_from_finish_brackets}{A logical indicating whether the final answer
should be extracted from the text inside the "FINISH\link{...}" brackets.}
}
\value{
A tidyprompt with the chain-of-thought prompt_wrap added.
}
\description{
Set chain of thought mode for a prompt
}
