% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_by_react.R
\name{answer_by_react}
\alias{answer_by_react}
\title{Set ReAct mode for a prompt}
\usage{
answer_by_react(prompt, extract_from_finish_brackets = TRUE)
}
\arguments{
\item{prompt}{A tidyprompt object or a single string}

\item{extract_from_finish_brackets}{A logical indicating whether the final answer
should be extracted from the text inside the "FINISH\link{...}" brackets.}
}
\value{
A tidyprompt with the ReAct-mode prompt_wrap added.
}
\description{
This function enables ReAct mode for the evaluation of a prompt.
In ReAct mode, the large language model (LLM) is asked to think step by step,
each time detailing a thought, action, and observation, to eventually arrive at
a final answer. It is hypothesized that this may increase LLM performance
at solving complex tasks. ReAct mode is inspired by the method described in
Yao et al. (2022).
}
\details{
Please note that ReAct mode may be most useful if in combination
with tools that the LLM can use. See, for example, 'add_tools()' for
enabling R function calling, or, for example, 'answer_as_code()' with
'output_as_tool = TRUE' for enabling R code evaluation as a tool.
}
\references{
Yao, S., Wu, Y., Cheung, W., Wang, Z., Narasimhan, K., & Kong, L. (2022).
ReAct: Synergizing Reasoning and Acting in Language Models.
\url{doi:10.48550/arXiv.2210.03629}
}
\seealso{
Other answer_by: 
\code{\link{answer_by_chain_of_thought}()}
}
\concept{answer_by}
