% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_prompt_wrap.R
\name{append_prompt_wrap}
\alias{append_prompt_wrap}
\alias{append_prompt_wrap.default}
\alias{append_prompt_wrap.prompt}
\title{Append a prompt wrap to a prompt object}
\usage{
append_prompt_wrap(prompt, prompt_wrap)

\method{append_prompt_wrap}{default}(prompt, prompt_wrap)

\method{append_prompt_wrap}{prompt}(prompt, prompt_wrap)
}
\arguments{
\item{prompt}{A prompt object to which the wrap is added.}

\item{prompt_wrap}{An additional structure to append to the prompt.}
}
\value{
A list combining the prompt and the appended wrap.

A list combining the prompt and the appended wrap.

A list combining the prompt and the appended wrap.
}
\description{
This function appends a prompt wrap (wrapper) to a prompt object. The method
can handle different types of input, such as a character or a prompt.
}
\section{Methods (by class)}{
\itemize{
\item \code{append_prompt_wrap(default)}: Default method for appending a prompt wrap.

This method handles general cases where the \code{prompt} is not of
class \code{prompt}. It attempts to convert the \code{prompt} to a 'prompt',
and then calls the \code{append_prompt_wrap} method for \code{prompt} objects.

\item \code{append_prompt_wrap(prompt)}: Method for appending a prompt wrap to a prompt.

}}
