% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_history.R
\name{chat_history.character}
\alias{chat_history.character}
\title{Method for \code{chat_history()} when the input is a single string}
\usage{
\method{chat_history}{character}(chat_history)
}
\arguments{
\item{chat_history}{A single string}
}
\value{
A valid chat history \code{data.frame} (of class \code{chat_history}),
with the 'role' set to 'user' and the 'content' set to the input string
}
\description{
Creates a \code{chat_history} object from a single string.
}
\keyword{internal}
