% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_history.R
\name{chat_history.data.frame}
\alias{chat_history.data.frame}
\title{Method for \code{chat_history()} when the input is a \code{data.frame}}
\usage{
\method{chat_history}{data.frame}(chat_history)
}
\arguments{
\item{chat_history}{A data frame with 'role' and 'content' columns,
where 'role' is either 'user', 'assistant', or 'system', and 'content' is a character string
representing a chat message}
}
\description{
Creates a \code{chat_history} object from a data frame.
}
\keyword{internal}
