% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_history.R
\name{chat_history.default}
\alias{chat_history.default}
\title{Default method for \code{chat_history()}}
\usage{
\method{chat_history}{default}(chat_history)
}
\arguments{
\item{chat_history}{Object which is not \code{character} or \code{data.frame}}
}
\description{
Calls error which indicates that the input was not a \code{character} or \code{data.frame}.
}
\details{
When input is a \code{character} or \code{data.frame}, the appropriate method will be called
(see `\code{\link[=chat_history.character]{chat_history.character()}} and \code{\link[=chat_history.data.frame]{chat_history.data.frame()}}).
}
\keyword{internal}
