% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{correct_prompt_list_order}
\alias{correct_prompt_list_order}
\title{Correct the order of a prompt list}
\usage{
correct_prompt_list_order(prompt_list)
}
\arguments{
\item{prompt_list}{A list of prompt_wrap objects}
}
\value{
A list of prompt_wrap objects with the order corrected.
'Mode' and 'toolset' prompts are placed at the bottom of the list.
}
\description{
Reorder prompt wrap objects in a prompt list so that "mode" and "toolset" prompts
are placed at the bottom, and the final prompt can be properly constructed.
}
