% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_provider.R
\name{create_google_gemini_llm_provider}
\alias{create_google_gemini_llm_provider}
\title{Create a new Google Gemini llm_provider instance}
\usage{
create_google_gemini_llm_provider(
  parameters = list(model = "gemini-1.5-flash", api_key =
    Sys.getenv("GOOGLE_AI_STUDIO_API_KEY"), base_url =
    "https://generativelanguage.googleapis.com/v1beta/models/")
)
}
\arguments{
\item{parameters}{A named list of parameters. Currently the following parameters are required:
\itemize{
\item model: The name of the model to use (e.g., "gemini-1.5-flash"; see https://ai.google.dev/gemini-api/docs/models/gemini)
\item api_key: The API key to use for authentication (get at https://aistudio.google.com/app/apikey)
\item base_url: The URL to the Google AI Studio API (default: "https://generativelanguage.googleapis.com/v1beta/models/")
Additional parameters are appended to the request body; see the Google AI Studio API
documentation for more information: https://ai.google.dev/gemini-api/docs/text-generation
and https://github.com/google/generative-ai-docs/blob/main/site/en/gemini-api/docs/get-started/rest.ipynb
}}
}
\value{
A new llm_provider object for use of the Google Gemini API
}
\description{
Create a new Google Gemini llm_provider instance
}
