% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_provider.R
\name{create_groq_llm_provider}
\alias{create_groq_llm_provider}
\title{Create a new Groq llm_provider instance}
\usage{
create_groq_llm_provider(
  parameters = list(model = "llama-3.1-8b-instant", api_key = Sys.getenv("GROQ_API_KEY"),
    url = "https://api.groq.com/openai/v1/chat/completions")
)
}
\arguments{
\item{parameters}{A named list of parameters. Currently the following parameters are required:
\itemize{
\item model: The name of the model to use (e.g., "llama-3.1-8b-instant")
\item api_key: The API key to use for authentication with the Groq API
\item url: The URL to the Groq API (default: "https://api.groq.com/openai/v1/chat/completions")
Additional parameters are appended to the request body; see the Groq API
documentation for more information: https://console.groq.com/docs/api-reference#chat-create
}}
}
\value{
A new llm_provider object for use of the Groq API
}
\description{
Create a new Groq llm_provider instance
}
