% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_feedback.R
\name{create_llm_feedback}
\alias{create_llm_feedback}
\title{Create an llm feedback object}
\usage{
create_llm_feedback(text)
}
\arguments{
\item{text}{A character string containing the feedback text. This will
be sent back to the LLM after not passing an extractor or validator function.}
}
\value{
An object of class "llm_feedback" containing the feedback text.
}
\description{
This object is used to send feedback to the LLM when a LLM reply
does not succesfully pass an extractor or validator function.
The extractor or validator function should then return this object
with the feedback text that should be sent to the LLM.
}
\examples{
create_llm_feedback("This is the feedback text")

is_integer <- function(x) {
 if (!is.integer(x)) {
   return(create_llm_feedback("The input is not an integer"))
 }
 return(TRUE)
}

}
