% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_provider.R, R/mistral_llm_provider.R
\name{create_mistral_llm_provider}
\alias{create_mistral_llm_provider}
\title{Create a new Mistral llm_provider instance}
\usage{
create_mistral_llm_provider(
  parameters = list(model = "mistral-small-latest", api_key =
    Sys.getenv("MISTRAL_API_KEY"), temperature = 0.3, top_p = 1, max_tokens = 0, stream =
    FALSE, stop = NULL, random_seed = 0, tool_choice = "auto", safe_prompt = FALSE)
)

create_mistral_llm_provider(
  parameters = list(model = "mistral-small-latest", api_key =
    Sys.getenv("MISTRAL_API_KEY"), temperature = 0.3, top_p = 1, max_tokens = 0, stream =
    FALSE, stop = NULL, random_seed = 0, tool_choice = "auto", safe_prompt = FALSE)
)
}
\arguments{
\item{parameters}{A named list of parameters. Currently the following parameters are required:
\itemize{
\item model: The name of the model to use (e.g., "mistral-small-latest")
\item api_key: The API key to use for authentication with the Mistral API. Store the API key safely.
\item temperature, top_p, max_tokens, stream, stop, random_seed, tool_choice, safe_prompt:
Additional options to control the model's behavior.
}}
}
\value{
A new llm_provider object for use of the Mistral API

A new llm_provider object for use of the Mistral API
}
\description{
This function creates a new llm_provider that interacts with the Mistral API.

This function creates a new llm_provider that interacts with the Mistral API.
}
