% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_provider.R
\name{create_ollama_llm_provider}
\alias{create_ollama_llm_provider}
\title{Create a new Ollama llm_provider instance}
\usage{
create_ollama_llm_provider(
  parameters = list(model = "llama3.1:8b", url = "http://localhost:11434/api/chat")
)
}
\arguments{
\item{parameters}{A named list of parameters. Currently the following parameters are required:
\itemize{
\item model: The name of the model to use (e.g., "llama3.1:8b")
\item url: The URL of the Ollama API endpoint
Additional parameters may be passed by adding them to the parameters list;
these parameters will be passed to the Ollama API via the body of the POST request.
Options specifically can be set with the $set_option function (e.g.,
ollama$set_option("option_name", "option_value")). See options at
https://ollama.com/docs/api/chat.
}}
}
\value{
A new llm_provider object for use of the Ollama API
}
\description{
Create a new Ollama llm_provider instance
}
