% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_provider.R
\name{create_openrouter_llm_provider}
\alias{create_openrouter_llm_provider}
\title{Create a new OpenRouter llm_provider instance}
\usage{
create_openrouter_llm_provider(
  parameters = list(model = "qwen/qwen-2.5-7b-instruct", api_key =
    Sys.getenv("OPENROUTER_API_KEY"), url =
    "https://openrouter.ai/api/v1/chat/completions"),
  verbose = getOption("tidyprompt.verbose", TRUE)
)
}
\arguments{
\item{parameters}{A named list of parameters. Currently the following parameters are required:
\itemize{
\item model: The name of the model to use (e.g., "qwen/qwen-2.5-7b-instruct"; see
https://openrouter.ai/docs/models).
\item api_key: The API key to use for authentication with the OpenRouter API (see
https://openrouter.ai/docs/api-keys).
\item url: The URL to the OpenRouter API (default: "https://openrouter.ai/api/v1/chat/completions").
Additional parameters are appended to the request body; see the OpenRouter API
documentation for more information: https://openrouter.ai/docs/parameters
}}

\item{verbose}{A logical indicating whether the interaction with the LLM provider
should be printed to the console. Default is TRUE.}
}
\value{
A new llm_provider object for use of the OpenRouter API
}
\description{
Create a new OpenRouter llm_provider instance
}
