% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_list.R
\name{create_prompt_list.character}
\alias{create_prompt_list.character}
\title{Method to create a prompt list object from a character string}
\usage{
\method{create_prompt_list}{character}(input, parameters = list())
}
\arguments{
\item{input}{Input to create_prompt_list; the base prompt}

\item{parameters}{Named list of parameters}
}
\value{
A prompt_list object
}
\description{
Method to create a prompt list object from a character string
}
