% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_list.R
\name{create_prompt_list.default}
\alias{create_prompt_list.default}
\title{Default method to create a prompt list object}
\usage{
\method{create_prompt_list}{default}(input, parameters = list())
}
\arguments{
\item{input}{Input to create_prompt_list}

\item{parameters}{Named list of parameters}
}
\value{
An error message stating that input type is not suitable
}
\description{
Is called when the input is not a character string or a prompt_list object.
}
