% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_list.R
\name{create_prompt_list}
\alias{create_prompt_list}
\title{Methods to create and manipulate prompt_list objects}
\usage{
create_prompt_list(input, parameters = list())
}
\arguments{
\item{input}{Input to create_prompt_list. If a prompt_list object is passed,
the parameters will be updated. If a character string is passed, a new prompt_list
object will be created with that character string as the base prompt.}

\item{parameters}{Named list of parameters to attach to the prompt_list object}
}
\value{
A prompt_list object (or an error if an unsuitable input is provided)
}
\description{
Methods to create and manipulate prompt_list objects
}
