% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{create_prompt_wrap}
\alias{create_prompt_wrap}
\title{Create a prompt wrap object}
\usage{
create_prompt_wrap(
  type = c("unspecified", "mode", "tool"),
  modify_fn = NULL,
  modify_fn_args = list(),
  validation_functions = list(),
  extraction_functions = list(),
  max_retries = 10
)
}
\arguments{
\item{type}{Type of the prompt; can be "unspecified", "mode", or "tool".
This is used to determine the order the prompt wrappers when constructing the final prompt text.
Mode prompts and toolset prompts are placed at the bottom of the prompt list (toolset after mode).}

\item{modify_fn}{Function that modifies the prompt text; takes two arguments: original_prompt_text and modify_fn_args.
This function will be applied to the previous prompt text in the prompt list.}

\item{modify_fn_args}{List of arguments to be passed to the modify_fn.}

\item{validation_functions}{List of functions that validate the (extracted) response to the prompt.
Should return TRUE on successful validation, or a 'llm_feedback' object upon failure.}

\item{extraction_functions}{List of functions that extract content from the response to the prompt.
Should return the extracted object on successful extraction, or a 'llm_feedback' object upon failure.}

\item{max_retries}{Maximum number of retries for this wrap's validation and extraction functions.}
}
\value{
A prompt wrap object
}
\description{
Create a prompt wrap object
}
