% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject_env_vars.R
\name{.inject_env_vars}
\alias{.inject_env_vars}
\title{Inject Environment Variables into a Function's Environment}
\usage{
.inject_env_vars(f, env = parent.frame())
}
\arguments{
\item{f}{A function whose environment will have variables injected from the
calling environment.}

\item{env}{The environment from which variables should be taken. Defaults to
the parent frame.}
}
\value{
The modified function with values from the environment injected into
its environment.
}
\description{
This is a helper function used internally within the package to inject values
from the environment into a function's environment. It captures all variables
from the specified environment (defaulting to the parent environment) and
binds them into the function's environment, so they can be accessed during
the function's execution.
}
\details{
This is particularly useful when you want to use non-standard evaluation
tools like \code{glue::glue()} within the function, and ensure that variables from
the calling environment are accessible inside the function.
}
\keyword{internal}
