% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_fn_json.R
\name{extraction_fn_json}
\alias{extraction_fn_json}
\title{Extract and parse JSONs from a string (LLM response)}
\usage{
extraction_fn_json(llm_response)
}
\arguments{
\item{llm_response}{A character string}
}
\value{
A list of parsed JSON objects
}
\description{
This function extracts JSON blocks from a string and parses them using \code{jsonlite::fromJSON()}.
This can be used to extract all JSONs from LLM responses, immediately converting
them to R objects.
}
\seealso{
Other text_helpers: 
\code{\link{df_to_string}()},
\code{\link{skim_with_labels_and_levels}()},
\code{\link{vector_list_to_string}()}
}
\concept{text_helpers}
