% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_using_tools.R
\name{gd_get_args_defaults_types}
\alias{gd_get_args_defaults_types}
\title{Get argument names, default values, and types from a function's formals}
\usage{
gd_get_args_defaults_types(func)
}
\arguments{
\item{func}{A function object}
}
\value{
A named list of arguments, where each argument is a list containing:
\itemize{
\item 'default_value': The default value of the argument. This is not
included if there is no default value. Note that when the default value
is included and is NULL, the default value is NULL (and not missing)
\item 'type': see 'gd_infer_type_from_default()'. If there is no default value,
the type is set to 'unknown'
@noRd
@keywords internal
}
}
\description{
This function extracts the argument names, default values, and types from
a function's formals. The types are inferred from the default values.
Is called internally to generate function documentation when there is none.
}
