% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{get_base_prompt}
\alias{get_base_prompt}
\title{Get base prompt from tidyprompt}
\usage{
get_base_prompt(tidyprompt)
}
\arguments{
\item{tidyprompt}{A tidyprompt object}
}
\value{
The base prompt from the tidyprompt
}
\description{
Get base prompt from tidyprompt
}
