% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{get_extractions_and_validations_from_prompt_list}
\alias{get_extractions_and_validations_from_prompt_list}
\title{Get the validation and extraction functions from a prompt list}
\usage{
get_extractions_and_validations_from_prompt_list(prompt_list)
}
\arguments{
\item{prompt_list}{A list of prompt_wrap objects}
}
\value{
A list with two elements: 'extractions' and 'validations'.#'
'extractions' is a list of extraction functions from the prompt list.#'
'validations' is a list of validation functions from the prompt list.
}
\description{
Get the validation and extraction functions from a prompt list
}
