% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{get_extractors_and_validators_from_prompt_list}
\alias{get_extractors_and_validators_from_prompt_list}
\title{Get the validator and extractor functions from a prompt list}
\usage{
get_extractors_and_validators_from_prompt_list(prompt_list)
}
\arguments{
\item{prompt_list}{A list of prompt_wrap objects}
}
\value{
A list with two elements: 'extractors' and 'validations'.#'
'extractors' is a list of extractor functions from the prompt list.#'
'validations' is a list of validation functions from the prompt list.
}
\description{
Get the validator and extractor functions from a prompt list
}
