% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{get_llm_provider_from_prompt_list}
\alias{get_llm_provider_from_prompt_list}
\title{Get the LLM provider from a prompt list}
\usage{
get_llm_provider_from_prompt_list(prompt_list)
}
\arguments{
\item{prompt_list}{A list of prompt_wrap objects}
}
\value{
The LLM provider object from the prompt list. The first
LLM provider found in the list will be returned (starting from the base prompt).
If no LLM provider is found in the list, NULL will be returned.
}
\description{
Get the LLM provider from a prompt list
}
