% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{get_prompt_wraps}
\alias{get_prompt_wraps}
\title{Get prompt wraps from tidyprompt}
\usage{
get_prompt_wraps(tidyprompt)
}
\arguments{
\item{tidyprompt}{A tidyprompt object}
}
\value{
A list of prompt wraps from the tidyprompt
}
\description{
Get prompt wraps from tidyprompt
}
\seealso{
Other tidyprompt_helpers: 
\code{\link{construct_prompt_text}()},
\code{\link{get_base_prompt}()},
\code{\link{get_extractions_and_validations}()},
\code{\link{get_prompt_wraps_ordered}()},
\code{\link{is_tidyprompt}()},
\code{\link{print.tidyprompt}()},
\code{\link{validate_tidyprompt}()}
}
\concept{tidyprompt_helpers}
