% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{get_prompt_wraps_ordered}
\alias{get_prompt_wraps_ordered}
\title{Extract only prompt wraps and reorder them in order of operations}
\usage{
get_prompt_wraps_ordered(tidyprompt)
}
\arguments{
\item{tidyprompt}{A tidyprompt object}
}
\value{
A list of prompt wraps from the tidyprompt, reordered in order of operations
}
\description{
This function extracts only prompt wraps from a tidyprompt object and reorders them.
The order of operations is as follows:
\enumerate{
\item "Unspecified"
\item "Mode"
\item "Tool"
}
}
\concept{tidyprompt_helpers}
