% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{is_prompt}
\alias{is_prompt}
\title{Check if object is a valid prompt}
\usage{
is_prompt(prompt)
}
\arguments{
\item{prompt}{A prompt object}
}
\value{
TRUE if the object is a valid prompt, otherwise FALSE
}
\description{
Check if object is a valid prompt
}
