% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{is_tidyprompt}
\alias{is_tidyprompt}
\title{Validate tidyprompt: returns TRUE if a valid tidyprompt object, otherwise FALSE}
\usage{
is_tidyprompt(tidyprompt)
}
\arguments{
\item{tidyprompt}{A tidyprompt object}
}
\value{
Logical indicating whether the input is a valid tidyprompt object
}
\description{
Validate tidyprompt: returns TRUE if a valid tidyprompt object, otherwise FALSE
}
\seealso{
Other tidyprompt: 
\code{\link{construct_prompt_text}()},
\code{\link{get_base_prompt}()},
\code{\link{get_prompt_wraps}()},
\code{\link{print.tidyprompt}()},
\code{\link{tidyprompt}()}
}
\concept{tidyprompt}
