% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_provider.R
\name{llm_provider}
\alias{llm_provider}
\title{Function to create llm_provider objects}
\usage{
llm_provider(
  complete_chat_function,
  parameters = list(),
  verbose = getOption("tidyprompt.verbose", TRUE),
  url = NULL,
  api_key = NULL
)
}
\arguments{
\item{complete_chat_function}{Function that will be called by the llm_provider
to complete a chat.

This function should take a chat_history dataframe as input (see ?chat_history),
and return a list of 'role' and 'content' for the next message (e.g., list(role = "user", content = "Hello")).

An llm_provider object will wrap the provided complete_chat_function with a validation of chat_history,
also turning a single string into a valid chat_history dataframe.
The provided function thus does not need to do this but should assume that
the input is a valid chat_history dataframe.

Parameters passed in the parameters argument may be accessed by the complete_chat_function;
this may be used to, for instance, store an API key, the name of the model to use, or other settings.}

\item{parameters}{A named list of parameters that will be attached to the
llm_provider object. These parameters can be used to configure the llm_provider.
For instance, they can be used to store a model's name, an API key, or an endpoint.
E.g., list(model = "my-llm-model", api_key = "my-api-key").}

\item{verbose}{A logical indicating whether the interaction with the LLM provider
should be printed to the console. Default is TRUE.}

\item{url}{The URL to the LLM provider API endpoint for chat completion}

\item{api_key}{The API key to use for authentication with the LLM provider API}
}
\value{
A new llm_provider object
}
\description{
This function can be used to create new llm_provider objects with different
implementations of the complete_chat function.
}
