% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_provider.R
\name{llm_provider_fake}
\alias{llm_provider_fake}
\title{Create a fake \code{\link[=llm_provider]{llm_provider()}} (for development and testing purposes)}
\usage{
llm_provider_fake(verbose = getOption("tidyprompt.verbose", TRUE))
}
\arguments{
\item{verbose}{A logical indicating whether the interaction with the \code{\link[=llm_provider]{llm_provider()}}
should be printed to the console. Default is TRUE.}
}
\value{
A new \code{\link[=llm_provider]{llm_provider()}} object for use of the fake LLM provider
}
\description{
This function creates a fake \code{\link[=llm_provider]{llm_provider()}} that can be used for development
and testing purposes. It is hardcoded to send back specific responses to
specific prompts that are used in vignettes, tests, and examples.
This is useful for running tests and builds in environments in which an
actual \code{\link[=llm_provider]{llm_provider()}} is not available.
}
