% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_provider.R
\name{llm_provider_groq}
\alias{llm_provider_groq}
\title{Create a new Groq llm_provider instance}
\usage{
llm_provider_groq(
  parameters = list(model = "llama-3.1-8b-instant", stream =
    getOption("tidyprompt.stream", TRUE)),
  verbose = getOption("tidyprompt.verbose", TRUE),
  url = "https://api.groq.com/openai/v1/chat/completions",
  api_key = Sys.getenv("GROQ_API_KEY")
)
}
\arguments{
\item{parameters}{A named list of parameters. Currently the following parameters are required:
\itemize{
\item model: The name of the model to use
\item stream: A logical indicating whether the API should stream responses
Additional parameters are appended to the request body; see the Groq API
documentation for more information: https://console.groq.com/docs/api-reference#chat-create
}}

\item{verbose}{A logical indicating whether the interaction with the LLM provider
should be printed to the console}

\item{url}{The URL to the Groq API}

\item{api_key}{The API key to use for authentication with the Groq API}
}
\value{
A new llm_provider object for use of the Groq API
}
\description{
Create a new Groq llm_provider instance
}
