% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{prompt.character}
\alias{prompt.character}
\title{Method to create a prompt object from a character string}
\usage{
\method{prompt}{character}(input)
}
\arguments{
\item{input}{Input to create_prompt; the base prompt}
}
\value{
A prompt object
}
\description{
Method to create a prompt object from a character string
}
