% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{prompt.default}
\alias{prompt.default}
\title{Default method to create a prompt object}
\usage{
\method{prompt}{default}(input)
}
\arguments{
\item{input}{Input to create_prompt}
}
\value{
An error message stating that input type is not suitable
}
\description{
Is called when the input is not a character string or a prompt object.
}
