% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_list.R
\name{prompt_list.character}
\alias{prompt_list.character}
\title{Method to create a prompt list object from a character string}
\usage{
\method{prompt_list}{character}(input)
}
\arguments{
\item{input}{Input to create_prompt_list; the base prompt}
}
\value{
A prompt_list object
}
\description{
Method to create a prompt list object from a character string
}
