% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_list.R
\name{prompt_list.prompt_list}
\alias{prompt_list.prompt_list}
\title{Method to create a prompt_list object from prompt_list}
\usage{
\method{prompt_list}{prompt_list}(input)
}
\arguments{
\item{input}{Input to create_prompt_list; the base prompt}
}
\value{
A prompt_list object
}
\description{
Method to create a prompt_list object from prompt_list
}
