% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_list.R
\name{prompt_list}
\alias{prompt_list}
\title{Methods to create and manipulate prompt_list objects}
\usage{
prompt_list(input)
}
\arguments{
\item{input}{Input to prompt_list. If a character string is passed,
a new prompt_list object will be created with that character string as the base prompt.}
}
\value{
A prompt_list object (or an error if an unsuitable input is provided)
}
\description{
Methods to create and manipulate prompt_list objects
}
