% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{prompt_wrap.default}
\alias{prompt_wrap.default}
\title{Default method for \code{\link[=prompt_wrap]{prompt_wrap()}}}
\usage{
\method{prompt_wrap}{default}(prompt, ...)
}
\arguments{
\item{prompt}{A single string or a \link{tidyprompt-class} object}

\item{...}{Additional arguments}
}
\value{
A \link{tidyprompt-class} object with the \code{\link[=prompt_wrap]{prompt_wrap()}} appended to it
}
\description{
Attempts to create a \link{tidyprompt-class} object from whatever is passed as 'prompt';
then calls the internal function to append the \code{\link[=prompt_wrap]{prompt_wrap()}}.
}
\keyword{internal}
