% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{prompt_wrap.prompt}
\alias{prompt_wrap.prompt}
\title{prompt_wrap method for when a prompt object is supplied}
\usage{
\method{prompt_wrap}{prompt}(prompt, ...)
}
\arguments{
\item{prompt}{A prompt object}

\item{...}{Additional arguments}
}
\value{
A prompt object with the new prompt wrap appended to it.
}
\description{
Calls the internal function to append the prompt wrap.
}
