% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{prompt_wrap_internal}
\alias{prompt_wrap_internal}
\title{Internal function to append a \code{\link[=prompt_wrap]{prompt_wrap()}} to a \code{\link[=tidyprompt]{tidyprompt()}} object}
\usage{
prompt_wrap_internal(
  prompt,
  modify_fn = NULL,
  extraction_fn = NULL,
  validation_fn = NULL,
  type = c("unspecified", "mode", "tool")
)
}
\arguments{
\item{prompt}{See \code{\link[=prompt_wrap]{prompt_wrap()}}}

\item{modify_fn}{See \code{\link[=prompt_wrap]{prompt_wrap()}}}

\item{extraction_fn}{See \code{\link[=prompt_wrap]{prompt_wrap()}}}

\item{validation_fn}{See \code{\link[=prompt_wrap]{prompt_wrap()}}}

\item{type}{See \code{\link[=prompt_wrap]{prompt_wrap()}}}
}
\value{
A \code{\link[=tidyprompt]{tidyprompt()}} object with the \code{\link[=prompt_wrap]{prompt_wrap()}} appended to it
}
\description{
Internal function to append a \code{\link[=prompt_wrap]{prompt_wrap()}} to a \code{\link[=tidyprompt]{tidyprompt()}} object
}
