% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_llm.R
\name{query_llm}
\alias{query_llm}
\title{...}
\usage{
query_llm(
  prompt,
  llm_provider,
  mode = c("default", "chain-of-thought"),
  validations = list(),
  max_retries = 10,
  verbose = global$settings$verbose_llm,
  cot_require_finish_brackets = TRUE,
  also_return_chat_history = FALSE
)
}
\arguments{
\item{prompt}{...}

\item{mode}{...}

\item{validations}{...}

\item{max_retries}{...}

\item{verbose}{...}

\item{cot_require_finish_brackets}{...}

\item{also_return_chat_history}{...}

\item{llm_model}{...}
}
\value{
...
}
\description{
...
}
\examples{
# ...
}
