% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_llm_provider.R
\name{set_llm_provider}
\alias{set_llm_provider}
\title{Set the llm_provider for a prompt}
\usage{
set_llm_provider(prompt, llm_provider)
}
\arguments{
\item{prompt}{A single character string or a prompt_list object}

\item{llm_provider}{An object of class llm_provider}
}
\value{
A prompt_list object with the llm_provider set as an attribute
}
\description{
Set the llm_provider for a prompt
}
