% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mode.R
\name{set_mode_chainofthought}
\alias{set_mode_chainofthought}
\title{Add chain-of-thought mode to a prompt}
\usage{
set_mode_chainofthought(
  prompt_wrap_or_list,
  extract_from_finish_brackets = TRUE
)
}
\arguments{
\item{prompt_wrap_or_list}{A single string, a prompt wrap,
or a list of prompt wraps.}

\item{extract_from_finish_brackets}{A logical indicating whether the final answer
should be extracted from the text inside the "FINISH\link{...}" brackets.}
}
\value{
A prompt list with the chain-of-thought mode added.
}
\description{
Add chain-of-thought mode to a prompt
}
