% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fn.R
\name{test_fn}
\alias{test_fn}
\title{Title}
\usage{
test_fn(
  no_arg,
  null_arg = NULL,
  call_arg = list.files(),
  string_arg = "string",
  number_arg = 123.45,
  integer_arg = 123,
  logical_arg = TRUE,
  vector_arg_numeric = c(1),
  vector_arg_character = c("a"),
  vector_arg_unknown = c(),
  vector_arg_mixed = c(1, "a", TRUE),
  vector_arg_match = c("a", "b", "c"),
  list_arg = list(1, 2, 3),
  named_list_arg = list(a = 1, b = 2, c = 3),
  nested_list_arg = list(list(1, 2), list(3, 4)),
  nested_named_list_arg = list(a = list(1, 2), b = list(3, 4))
)
}
\arguments{
\item{no_arg}{No default}

\item{null_arg}{Default NULL}

\item{call_arg}{Call}

\item{string_arg}{String}

\item{number_arg}{Number}

\item{integer_arg}{Integer}

\item{logical_arg}{Logical}

\item{vector_arg_numeric}{Numeric vector}

\item{vector_arg_character}{Character vector}

\item{vector_arg_unknown}{Unknown vector}

\item{vector_arg_mixed}{Mixed vector}

\item{vector_arg_match}{Match vector}

\item{list_arg}{List}

\item{named_list_arg}{Named list}

\item{nested_list_arg}{Nested list}

\item{nested_named_list_arg}{Nested named list}
}
\value{
...
}
\description{
Description
}
