% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt-package.R
\docType{package}
\name{tidyprompt-package}
\alias{tidyprompt-package}
\title{tidyprompt: Empower your LLM, the tidy way}
\description{
The `tidyprompt` package allows users to construct LLM prompts using tidyverse-inspired piping syntax, with a library of prompt wrappers or custom ones. It supports structured output extraction and validation, with automatic retries if necessary. Additionally, it enables specific LLM reasoning modes, autonomous R function calling ('tools') for LLMs, and compatibility with multiple LLM providers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjarkvandemerwe/tidyprompt}
  \item Report bugs at \url{https://github.com/tjarkvandemerwe/tidyprompt/issues}
}

}
\author{
\strong{Maintainer}: Luka Koning \email{l.koning@kennispunttwente.nl}

Authors:
\itemize{
  \item Tjark Van de Merwe \email{t.vandemerwe@kennispunttwente.nl}
}

}
\keyword{internal}
