% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{tidyprompt.character}
\alias{tidyprompt.character}
\title{Method to create a tidyprompt object from a character string}
\usage{
\method{tidyprompt}{character}(input)
}
\arguments{
\item{input}{Input to create_tidyprompt; the base prompt}
}
\value{
A tidyprompt object
}
\description{
Method to create a tidyprompt object from a character string
}
\keyword{internal}
