% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{tidyprompt.default}
\alias{tidyprompt.default}
\title{Default method to create a tidyprompt object}
\usage{
\method{tidyprompt}{default}(input)
}
\arguments{
\item{input}{Input to create_tidyprompt}
}
\value{
An error message stating that input type is not suitable
}
\description{
Is called when the input is not a character string or a tidyprompt object.
}
\keyword{internal}
