% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{tidyprompt}
\alias{tidyprompt}
\title{Methods to create and manipulate prompt objects}
\usage{
tidyprompt(input)
}
\arguments{
\item{input}{Input to prompt. If a character string is passed,
a new prompt object will be created with that character string as the base prompt.}
}
\value{
A prompt object (or an error if an unsuitable input is provided)
}
\description{
Methods to create and manipulate prompt objects
}
