% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{tidyprompt.tidyprompt}
\alias{tidyprompt.tidyprompt}
\title{Validate tidyprompt}
\usage{
\method{tidyprompt}{tidyprompt}(input)
}
\arguments{
\item{input}{A tidyprompt object}
}
\value{
the input object if the tidyprompt is valid,
otherwise an error is thrown
}
\description{
Validate tidyprompt
}
\keyword{internal}
