% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{tool_extraction}
\alias{tool_extraction}
\title{Extraction function for LLM function calling}
\usage{
tool_extraction(llm_response, tool_functions)
}
\arguments{
\item{llm_response}{LLM response text to parse for a function call}

\item{tool_functions}{List of R functions that can be called by the LLM;
should have internal docstring-like documentation with 'llm_tool::' tags}
}
\value{
If a tool is called: a character string with the result of the function call or an error message.
If no tool is called: the original LLM response.
}
\description{
Extraction function for LLM function calling
}
