% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{validate_prompt}
\alias{validate_prompt}
\title{Validate prompt}
\usage{
validate_prompt(prompt)
}
\arguments{
\item{prompt}{A prompt object}
}
\value{
TRUE if the prompt is valid, otherwise an error is thrown
}
\description{
Validate prompt
}
