% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{validate_prompt_list}
\alias{validate_prompt_list}
\title{Validate a prompt list}
\usage{
validate_prompt_list(prompt_wrap_or_list)
}
\arguments{
\item{prompt_wrap_or_list}{A single string, a prompt_wrap object, or a list
of prompt_wrap objects.}
}
\value{
A valid prompt list. If the input contains errors and cannot be turned
into a valid prompt list, an error will be thrown.
}
\description{
Tidyprompt will use a list of prompt_wrap objects to organise a base prompt and
modifications to that base prompt, plus extractor and validation functions which
need to be applied after the prompt is sent to the LLM. An LLM provider,
to be used for evaluation of the final prompt, can also be kept within one of
the prompt wrap objects within the prompt list.
}
