% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{validate_tidyprompt}
\alias{validate_tidyprompt}
\title{Validate tidyprompt, returning self if valid, otherwise error}
\usage{
validate_tidyprompt(tidyprompt)
}
\arguments{
\item{tidyprompt}{A tidyprompt object}
}
\value{
TRUE if the tidyprompt is valid, otherwise an error is thrown
}
\description{
Validate tidyprompt, returning self if valid, otherwise error
}
\seealso{
Other tidyprompt: 
\code{\link{construct_prompt_text}()},
\code{\link{get_base_prompt}()},
\code{\link{get_prompt_wraps}()},
\code{\link{is_tidyprompt}()},
\code{\link{print.tidyprompt}()},
\code{\link{tidyprompt}()}
}
\concept{tidyprompt}
