% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{add_tools}
\alias{add_tools}
\title{Add function-calling to prompt}
\usage{
add_tools(prompt, tool_functions = list())
}
\arguments{
\item{prompt}{...}

\item{tool_functions}{...}
}
\value{
...
}
\description{
Add function-calling to prompt
}
