% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_as_boolean.R
\name{answer_as_boolean}
\alias{answer_as_boolean}
\title{Make LLM answer as a boolean (TRUE or FALSE)}
\usage{
answer_as_boolean(prompt, add_instruction_to_prompt = TRUE)
}
\arguments{
\item{prompt}{A character string or a tidyprompt object}

\item{add_instruction_to_prompt}{(optional) Add instruction for replying
as a boolean to the prompt text. Set to FALSE for debugging if extractions/validations
are working as expected (without instruction the answer should fail the
validation function, initiating a retry).}
}
\value{
A tidyprompt with an added prompt wrapper object which
will ensure that the LLM response is a boolean.
}
\description{
Make LLM answer as a boolean (TRUE or FALSE)
}
\seealso{
Other answer_as: 
\code{\link{answer_as_code}()},
\code{\link{answer_as_integer}()},
\code{\link{answer_as_list}()},
\code{\link{answer_as_named_list}()},
\code{\link{answer_as_regex}()}
}
\concept{answer_as}
