% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_as_integer.R
\name{answer_as_integer}
\alias{answer_as_integer}
\title{Make LLM answer as an integer (between min and max)}
\usage{
answer_as_integer(
  prompt,
  min = NULL,
  max = NULL,
  add_instruction_to_prompt = TRUE
)
}
\arguments{
\item{prompt}{A single string, a prompt_wrap object, or a list
of prompt_wrap objects.}

\item{min}{(optional) Minimum value for the integer}

\item{max}{(optional) Maximum value for the integer}

\item{add_instruction_to_prompt}{(optional) Add instruction for replying
as an integer to the prompt text. Set to FALSE for debugging if extractions/validations
are working as expected (without instruction the answer should fail the
validation function, initiating a retry).}
}
\value{
A prompt list with an added prompt wrapper object which
will ensure that the LLM response is an integer.
}
\description{
Make LLM answer as an integer (between min and max)
}
