% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_as_list.R
\name{answer_as_list}
\alias{answer_as_list}
\title{Make LLM answer as a list of items}
\usage{
answer_as_list(
  prompt,
  item_name = "item",
  item_explanation = NULL,
  n_unique_items = NULL
)
}
\arguments{
\item{prompt}{A single string or a \code{\link[=tidyprompt]{tidyprompt()}} object}

\item{item_name}{(optional) Name of the items in the list}

\item{item_explanation}{(optional) Additional explanation of what an item
should be. Item explanation should be a single string. It will be
appended after the list instruction}

\item{n_unique_items}{(optional) Number of unique items required in the list}
}
\value{
A \code{\link[=tidyprompt]{tidyprompt()}} with an added \code{\link[=prompt_wrap]{prompt_wrap()}} which
will ensure that the LLM response is a list of items
}
\description{
Make LLM answer as a list of items
}
\examples{
\dontrun{
  "What are some delicious fruits?" |>
    answer_as_list(item_name = "fruit", n_unique_items = 5) |>
    send_prompt(llm_provider_ollama())
#   --- Sending request to LLM provider (llama3.1:8b): ---
#     What are some delicious fruits?
#
#     Respond with a list, like so:
#       -- <<fruit 1>>
#       -- <<fruit 2>>
#       etc.
#     The list should contain 5 unique items.
#   --- Receiving response from LLM provider: ---
#     Here's a list of delicious fruits:
#
#     -- Strawberries
#     -- Pineapples
#     -- Mangoes
#     -- Papayas
#     -- Kiwis
# [1] "Strawberries" "Pineapples"   "Mangoes"      "Papayas"      "Kiwis"
}

}
\seealso{
\code{\link[=answer_as_named_list]{answer_as_named_list()}}

Other answer_as: 
\code{\link{answer_as_boolean}()},
\code{\link{answer_as_code}()},
\code{\link{answer_as_integer}()},
\code{\link{answer_as_named_list}()},
\code{\link{answer_as_regex}()}
}
\concept{answer_as}
