% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_as_named_list.R
\name{answer_as_named_list}
\alias{answer_as_named_list}
\title{Extract named list from LLM response with optional item instructions and validations}
\usage{
answer_as_named_list(
  prompt,
  item_names,
  item_instructions = NULL,
  item_validations = NULL
)
}
\arguments{
\item{prompt}{A single string or a \code{\link[=tidyprompt]{tidyprompt()}} object}

\item{item_names}{A character vector specifying the expected item names}

\item{item_instructions}{An optional named list of additional instructions for each item}

\item{item_validations}{An optional named list of validation functions for each item.
Like validation functions for a \code{\link[=prompt_wrap]{prompt_wrap()}}, these functions should return
\code{\link[=llm_feedback]{llm_feedback()}} if the validation fails. If the validation
is successful, the function should return TRUE}
}
\value{
A \code{\link[=tidyprompt]{tidyprompt()}} with an added \code{\link[=prompt_wrap]{prompt_wrap()}} that ensures
the LLM response is a named list with the specified item names, optional
instructions, and validations.
}
\description{
Extract named list from LLM response with optional item instructions and validations
}
\examples{
\dontrun{
  persona <- "Create a persona for me, please." |>
    answer_as_named_list(
      item_names = c("name", "age", "occupation"),
      item_instructions = list(
        name = "The name of the persona",
        age = "The age of the persona",
        occupation = "The occupation of the persona"
      )
    ) |> send_prompt(llm_provider_ollama())
  # --- Sending request to LLM provider (llama3.1:8b): ---
  #   Create a persona for me, please.
  #
  #   Respond with a named list like so:
  #     -- name: <<value>> (The name of the persona)
  #     -- age: <<value>> (The age of the persona)
  #     -- occupation: <<value>> (The occupation of the persona)
  #   Each name must correspond to: name, age, occupation
  # --- Receiving response from LLM provider: ---
  #   Here is your persona:
  #
  #   -- name: Astrid Welles
  #   -- age: 32
  #   -- occupation: Museum Curator
  persona$name
  # [1] "Astrid Welles"
  persona$age
  # [1] "32"
  persona$occupation
  # [1] "Museum Curator"
}
}
\seealso{
\code{\link[=answer_as_list]{answer_as_list()}} \code{\link[=llm_feedback]{llm_feedback()}}

Other pre_built_prompt_wraps: 
\code{\link{add_text}()},
\code{\link{add_tools}()},
\code{\link{answer_as_boolean}()},
\code{\link{answer_as_code}()},
\code{\link{answer_as_integer}()},
\code{\link{answer_as_list}()},
\code{\link{answer_as_regex}()},
\code{\link{answer_by_chain_of_thought}()},
\code{\link{answer_by_react}()},
\code{\link{prompt_wrap}()},
\code{\link{quit_if}()},
\code{\link{set_system_prompt}()}

Other answer_as_prompt_wraps: 
\code{\link{answer_as_boolean}()},
\code{\link{answer_as_code}()},
\code{\link{answer_as_integer}()},
\code{\link{answer_as_list}()},
\code{\link{answer_as_regex}()}
}
\concept{answer_as_prompt_wraps}
\concept{pre_built_prompt_wraps}
