% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_as_regex.R
\name{answer_as_regex}
\alias{answer_as_regex}
\title{Make LLM answer match a specific regex}
\usage{
answer_as_regex(prompt, regex)
}
\arguments{
\item{prompt}{A single string or a \code{\link[=tidyprompt]{tidyprompt()}} object}

\item{regex}{A character string specifying the regular expression the response must match}
}
\value{
A \code{\link[=tidyprompt]{tidyprompt()}} with an added \code{\link[=prompt_wrap]{prompt_wrap()}} which will ensure
that the LLM response matches the specified regex
}
\description{
Make LLM answer match a specific regex
}
\seealso{
Other answer_as: 
\code{\link{answer_as_boolean}()},
\code{\link{answer_as_code}()},
\code{\link{answer_as_integer}()},
\code{\link{answer_as_list}()},
\code{\link{answer_as_named_list}()}
}
\concept{answer_as}
