% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_by_chain_of_thought.R
\name{answer_by_chain_of_thought}
\alias{answer_by_chain_of_thought}
\title{Set chain of thought mode for a prompt}
\usage{
answer_by_chain_of_thought(prompt, extract_from_finish_brackets = TRUE)
}
\arguments{
\item{prompt}{A tidyprompt object or a single string}

\item{extract_from_finish_brackets}{A logical indicating whether the final answer
should be extracted from the text inside the "FINISH\link{...}" brackets.}
}
\value{
A tidyprompt with the chain-of-thought prompt_wrap added.
}
\description{
This function enables chain of thought mode for evaluation of a prompt.
In chain of thought mode, the large language model (LLM) is asked to think
step by step to arrive at a final answer. It is hypothesized that this may
increase LLM performance at solving complex tasks. Chain of thought mode
is inspired by the method described in Wei et al. (2022).
}
\references{
Wei, J., Wang, X., Schuurmans, D., Bosma, M., Ichter, B., Xia, F., Chi, E., Le, Q., & Zhou, D. (2022).
Chain-of-Thought Prompting Elicits Reasoning in Large Language Models.
\url{doi:10.48550/arXiv.2201.11903}
}
\seealso{
Other answer_by: 
\code{\link{answer_by_react}()}
}
\concept{answer_by}
