% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_by_react.R
\name{answer_by_react}
\alias{answer_by_react}
\title{Set ReAct mode for a prompt}
\usage{
answer_by_react(prompt, extract_from_finish_brackets = TRUE)
}
\arguments{
\item{prompt}{A tidyprompt object or a single string}

\item{extract_from_finish_brackets}{A logical indicating whether the final answer
should be extracted from the text inside the "FINISH\link{...}" brackets.}
}
\value{
A tidyprompt with the ReAct-mode prompt_wrap added.
}
\description{
Set ReAct mode for a prompt
}
\details{
Please note that ReAct mode may be most useful if in combination
with tools that the LLM can use. See, for example, 'add_tools()' for
enabling R function calling, or, for example, 'answer_as_code()' with
'output_as_tool = TRUE' for enabling R code evaluation as a tool.
}
