% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_history.R
\name{chat_history.character}
\alias{chat_history.character}
\title{Method for \code{chat_history()} when the input is a single string}
\usage{
\method{chat_history}{character}(chat_history)
}
\arguments{
\item{chat_history}{A single string}
}
\description{
Creates a \code{chat_history} object from a single string.
}
