% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_history.R
\name{chat_history}
\alias{chat_history}
\title{Create or validate \code{chat_history} object}
\usage{
chat_history(chat_history)
}
\arguments{
\item{chat_history}{A single string, a dataframe with 'role' and 'content' columns,
or NULL.}
}
\value{
A valid chat history dataframe (of class `chat_history``).
}
\description{
This function creates and validates a \code{chat_history} object, ensuring that it matches
the expected format with 'role' and 'content' columns. It has separate methods
for \code{data.frame} and \code{character} inputs and includes a helper function to add a
system prompt to the chat history.
}
\concept{chat_history}
