% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_prompt_text.R
\name{construct_prompt_text}
\alias{construct_prompt_text}
\title{Construct the prompt text}
\usage{
construct_prompt_text(prompt)
}
\arguments{
\item{prompt}{A single character string,
or a prompt_list object.}
}
\value{
The constructed prompt text as a character string.
The prompt text is constructed by applying the modifying functions
of all prompt wrappers in the prompt list to the base prompt text.
}
\description{
Construct the prompt text
}
