% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{construct_prompt_text}
\alias{construct_prompt_text}
\title{Construct prompt text from a Tidyprompt object}
\usage{
construct_prompt_text(x, llm_provider = NULL)
}
\arguments{
\item{x}{A Tidyprompt object.}

\item{llm_provider}{An optional LLM provider object.}
}
\value{
The constructed prompt text.
}
\description{
Construct prompt text from a Tidyprompt object
}
\seealso{
Other tidyprompt: 
\code{\link{get_prompt_wraps}()},
\code{\link{is_tidyprompt}()},
\code{\link{tidyprompt}()},
\code{\link{tidyprompt-class}}
}
\concept{tidyprompt}
