% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{construct_prompt_text}
\alias{construct_prompt_text}
\title{Construct prompt text from a tidyprompt object}
\usage{
construct_prompt_text(tidyprompt)
}
\arguments{
\item{tidyprompt}{A tidyprompt object}
}
\value{
The prompt text constructed from the tidyprompt object
}
\description{
Construct prompt text from a tidyprompt object
}
\seealso{
Other tidyprompt: 
\code{\link{get_base_prompt}()},
\code{\link{get_prompt_wraps}()},
\code{\link{is_tidyprompt}()},
\code{\link{print.tidyprompt}()},
\code{\link{tidyprompt}()}
}
\concept{tidyprompt}
