% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{construct_prompt_text}
\alias{construct_prompt_text}
\title{Construct prompt text from a prompt object}
\usage{
construct_prompt_text(prompt)
}
\arguments{
\item{prompt}{A prompt object}
}
\value{
The prompt text constructed from the prompt object
}
\description{
Construct prompt text from a prompt object
}
