% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{construct_prompt_text}
\alias{construct_prompt_text}
\title{Construct a final prompt text from a prompt list}
\usage{
construct_prompt_text(prompt_wrap_or_list)
}
\arguments{
\item{prompt_wrap_or_list}{A single string, a prompt_wrap object, or a list
of prompt_wrap objects.}
}
\value{
A character string containing the final prompt text.
}
\description{
Construct a final prompt text from a prompt list
}
