% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_provider.R
\name{create_ollama_llm_provider}
\alias{create_ollama_llm_provider}
\title{Create a new Ollama llm_provider instance}
\usage{
create_ollama_llm_provider(
  parameters = list(model = "llama3.1:8b", url = "http://localhost:11434/api/chat")
)
}
\arguments{
\item{parameters}{A named list of parameters. Currently the following parameters are required:
\itemize{
\item model: The name of the model to use (e.g., "llama3.1:8b")
\item url: The URL of the Ollama API endpoint
}}
}
\value{
A new llm_provider object for use of the Ollama API
}
\description{
Create a new Ollama llm_provider instance
}
