% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{get_base_prompt}
\alias{get_base_prompt}
\title{Get base prompt from tidyprompt}
\usage{
get_base_prompt(tidyprompt)
}
\arguments{
\item{tidyprompt}{A tidyprompt object}
}
\value{
The base prompt from the tidyprompt
}
\description{
Get base prompt from tidyprompt
}
\seealso{
Other tidyprompt: 
\code{\link{construct_prompt_text}()},
\code{\link{get_extractions_and_validations}()},
\code{\link{get_prompt_wraps}()},
\code{\link{is_tidyprompt}()},
\code{\link{tidyprompt}()},
\code{\link{validate_tidyprompt}()}
}
\concept{tidyprompt}
