% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{get_base_prompt}
\alias{get_base_prompt}
\title{Get base prompt from prompt}
\usage{
get_base_prompt(prompt)
}
\arguments{
\item{prompt}{A prompt object}
}
\value{
The base prompt from the prompt
}
\description{
Get base prompt from prompt
}
