% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{get_extractions_and_validations}
\alias{get_extractions_and_validations}
\title{Get extractions and validations from a tidyprompt}
\usage{
get_extractions_and_validations(tidyprompt)
}
\arguments{
\item{tidyprompt}{A tidyprompt object}
}
\value{
A list with two lists: extractions and validations
}
\description{
Get extractions and validations from a tidyprompt
}
