% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{get_extractions_and_validations}
\alias{get_extractions_and_validations}
\title{Get extractions and validations from a prompt}
\usage{
get_extractions_and_validations(prompt)
}
\arguments{
\item{prompt}{A prompt object}
}
\value{
A list with two lists: extractions and validations
}
\description{
Get extractions and validations from a prompt
}
