% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{get_prompt_wraps}
\alias{get_prompt_wraps}
\title{Get prompt wraps from a Tidyprompt}
\usage{
get_prompt_wraps(x, order = c("default", "modification", "evaluation"))
}
\arguments{
\item{x}{A Tidyprompt object.}

\item{order}{The order in which to return the prompt wraps. Options: "default", "modification", "evaluation".}
}
\value{
A list of prompt wraps.
}
\description{
Get prompt wraps from a Tidyprompt
}
\seealso{
Other tidyprompt: 
\code{\link{construct_prompt_text}()},
\code{\link{is_tidyprompt}()},
\code{\link{tidyprompt}()},
\code{\link{tidyprompt-class}}
}
\concept{tidyprompt}
