% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{get_prompt_wraps}
\alias{get_prompt_wraps}
\title{Get prompt wraps from tidyprompt}
\usage{
get_prompt_wraps(tidyprompt)
}
\arguments{
\item{tidyprompt}{A tidyprompt object}
}
\value{
A list of prompt wraps from the tidyprompt
}
\description{
Get prompt wraps from tidyprompt
}
