% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{get_prompt_wraps}
\alias{get_prompt_wraps}
\title{Get prompt wraps from prompt}
\usage{
get_prompt_wraps(prompt)
}
\arguments{
\item{prompt}{A prompt object}
}
\value{
A list of prompt wraps from the prompt
}
\description{
Get prompt wraps from prompt
}
