% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{is_tidyprompt}
\alias{is_tidyprompt}
\title{Check if object is a valid tidyprompt}
\usage{
is_tidyprompt(tidyprompt)
}
\arguments{
\item{tidyprompt}{A tidyprompt object}
}
\value{
TRUE if the object is a valid tidyprompt, otherwise FALSE
}
\description{
Check if object is a valid tidyprompt
}
\seealso{
Other tidyprompt_helpers: 
\code{\link{construct_prompt_text}()},
\code{\link{get_base_prompt}()},
\code{\link{get_extractions_and_validations}()},
\code{\link{get_prompt_wraps}()},
\code{\link{get_prompt_wraps_ordered}()},
\code{\link{print.tidyprompt}()},
\code{\link{validate_tidyprompt}()}
}
\concept{tidyprompt_helpers}
