% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{is_tidyprompt}
\alias{is_tidyprompt}
\title{Check if object is a valid tidyprompt}
\usage{
is_tidyprompt(tidyprompt)
}
\arguments{
\item{tidyprompt}{A tidyprompt object}
}
\value{
TRUE if the object is a valid tidyprompt, otherwise FALSE
}
\description{
Check if object is a valid tidyprompt
}
