% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{is_tidyprompt}
\alias{is_tidyprompt}
\title{Validate tidyprompt, returning TRUE if valid, otherwise FALSE}
\usage{
is_tidyprompt(tidyprompt)
}
\arguments{
\item{tidyprompt}{A tidyprompt object}
}
\value{
TRUE if the object is a valid tidyprompt, otherwise FALSE
}
\description{
Validate tidyprompt, returning TRUE if valid, otherwise FALSE
}
\seealso{
Other tidyprompt: 
\code{\link{construct_prompt_text}()},
\code{\link{get_base_prompt}()},
\code{\link{get_prompt_wraps}()},
\code{\link{print.tidyprompt}()},
\code{\link{tidyprompt}()},
\code{\link{validate_tidyprompt}()}
}
\concept{tidyprompt}
