% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{is_tidyprompt}
\alias{is_tidyprompt}
\title{Check if object is a Tidyprompt}
\usage{
is_tidyprompt(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
TRUE if the object is a valid Tidyprompt, otherwise FALSE
}
\description{
Check if object is a Tidyprompt
}
\seealso{
Other tidyprompt: 
\code{\link{construct_prompt_text}()},
\code{\link{get_prompt_wraps}()},
\code{\link{tidyprompt}()},
\code{\link{tidyprompt-class}}
}
\concept{tidyprompt}
