% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_provider.R
\name{llm_provider_google_gemini}
\alias{llm_provider_google_gemini}
\title{Create a new Google Gemini \code{\link[=llm_provider]{llm_provider()}} instance}
\usage{
llm_provider_google_gemini(
  parameters = list(model = "gemini-1.5-flash"),
  verbose = getOption("tidyprompt.verbose", TRUE),
  url = "https://generativelanguage.googleapis.com/v1beta/models/",
  api_key = Sys.getenv("GOOGLE_AI_STUDIO_API_KEY")
)
}
\arguments{
\item{parameters}{A named list of parameters. Currently the following parameters are required:
\itemize{
\item model: The name of the model to use (see: https://ai.google.dev/gemini-api/docs/models/gemini)
Additional parameters are appended to the request body; see the Google AI Studio API
documentation for more information: https://ai.google.dev/gemini-api/docs/text-generation
& https://github.com/google/generative-ai-docs/blob/main/site/en/gemini-api/docs/get-started/rest.ipynb
}}

\item{verbose}{A logical indicating whether the interaction with the LLM provider
should be printed to the console}

\item{url}{The URL to the Google Gemini API}

\item{api_key}{The API key to use for authentication with the Google Gemini API
(see: https://aistudio.google.com/app/apikey)}
}
\value{
A new \code{\link[=llm_provider]{llm_provider()}} object for use of the Google Gemini API
}
\description{
Creates an \code{\link[=llm_provider]{llm_provider()}} object that interacts with the Google Gemini API.
Streaming is not yet supported in this implementation.
}
\seealso{
Other llm_provider: 
\code{\link{llm_provider}},
\code{\link{llm_provider_fake}()},
\code{\link{llm_provider_groq}()},
\code{\link{llm_provider_mistral}()},
\code{\link{llm_provider_ollama}()},
\code{\link{llm_provider_openai}()},
\code{\link{llm_provider_openrouter}()},
\code{\link{llm_provider_xai}()},
\code{\link{make_llm_provider_request}()}
}
\concept{llm_provider}
