% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_provider.R, R/llm_provider_R6.R
\name{llm_provider_openai}
\alias{llm_provider_openai}
\title{Create a new OpenAI llm_provider instance}
\usage{
llm_provider_openai(
  parameters = list(model = "gpt-4o-mini", stream = TRUE),
  verbose = getOption("tidyprompt.verbose", TRUE),
  url = "https://api.openai.com/v1/chat/completions",
  api_key = Sys.getenv("OPENAI_API_KEY")
)

llm_provider_openai(
  parameters = list(model = "gpt-4o-mini", stream = TRUE),
  verbose = getOption("tidyprompt.verbose", TRUE),
  url = "https://api.openai.com/v1/chat/completions",
  api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{parameters}{A named list of parameters. Currently the following parameters are required:
\itemize{
\item model: The name of the model to use
\item api_key: The API key to use for authentication with the OpenAI API. This
should be a project API key (not a user API key)
\item url: The URL to the OpenAI API (may also be an alternative endpoint
that provides a similar API.)
\item stream: A logical indicating whether the API should stream responses
Additional parameters are appended to the request body; see the OpenAI API
documentation for more information: https://platform.openai.com/docs/api-reference/chat
}}

\item{verbose}{A logical indicating whether the interaction with the LLM provider
should be printed to the console. Default is TRUE.}

\item{url}{The URL to the OpenAI API}

\item{api_key}{The API key to use for authentication with the OpenAI API}
}
\value{
A new llm_provider object for use of the OpenAI API

A new llm_provider object for use of the OpenAI API
}
\description{
This function creates a new llm_provider that interacts with the Open AI API

This function creates a new llm_provider that interacts with the Open AI API
}
