% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_provider.R
\name{llm_provider_xai}
\alias{llm_provider_xai}
\title{Create a new XAI (Grok) llm_provider instance}
\usage{
llm_provider_xai(
  parameters = list(model = "grok-beta", stream = getOption("tidyprompt.stream", TRUE)),
  verbose = getOption("tidyprompt.verbose", TRUE),
  url = "https://api.x.ai/v1/chat/completions",
  api_key = Sys.getenv("XAI_API_KEY")
)
}
\arguments{
\item{parameters}{A named list of parameters. Currently the following parameters are required:
\itemize{
\item model: The name of the model to use
\item stream: A logical indicating whether the API should stream responses
Additional parameters are appended to the request body; see the XAI API
documentation for more information: https://docs.x.ai/api/endpoints#chat-completions
}}

\item{verbose}{A logical indicating whether the interaction with the LLM provider
should be printed to the console. Default is TRUE.}

\item{url}{The URL to the XAI API}

\item{api_key}{The API key to use for authentication with the XAI API}
}
\value{
A new llm_provider object for use of the XAI API
}
\description{
Create a new XAI (Grok) llm_provider instance
}
