% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistent_chat.R
\name{persistent_chat-class}
\alias{persistent_chat-class}
\title{PersistentChat R6 class}
\description{
A class for managing a persistent chat with a large language model (LLM).

While 'tidyprompt' is primariy focused on automatic interactions with
LLMs through \code{\link[=send_prompt]{send_prompt()}} using a \linkS4class{tidyprompt} object with
\code{\link[=prompt_wrap]{prompt_wrap()}}, this class may be useful for having a manual conversation
with an LLM. (It may specifically be used to continue a chat history which was
returned by \code{\link[=send_prompt]{send_prompt()}} with \code{return_mode = "full"}.)
}
\seealso{
\linkS4class{llm_provider} \code{\link[=chat_history]{chat_history()}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{chat_history}}{A \code{\link[=chat_history]{chat_history()}} object}

\item{\code{llm_provider}}{A \linkS4class{llm_provider} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PersistentChat-new}{\code{persistent_chat-class$new()}}
\item \href{#method-PersistentChat-chat}{\code{persistent_chat-class$chat()}}
\item \href{#method-PersistentChat-reset_chat_history}{\code{persistent_chat-class$reset_chat_history()}}
\item \href{#method-PersistentChat-clone}{\code{persistent_chat-class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersistentChat-new"></a>}}
\if{latex}{\out{\hypertarget{method-PersistentChat-new}{}}}
\subsection{Method \code{new()}}{
Initialize the PersistentChat object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persistent_chat-class$new(llm_provider, chat_history = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{llm_provider}}{A \linkS4class{llm_provider} object}

\item{\code{chat_history}}{(optional) A \code{\link[=chat_history]{chat_history()}} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The initialized PersistentChat object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersistentChat-chat"></a>}}
\if{latex}{\out{\hypertarget{method-PersistentChat-chat}{}}}
\subsection{Method \code{chat()}}{
Add a message to the chat history and get a response from the LLM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persistent_chat-class$chat(msg, role = "user", verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{Message to add to the chat history}

\item{\code{role}}{Role of the message}

\item{\code{verbose}}{Whether to print the interaction to the console}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The response from the LLM
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersistentChat-reset_chat_history"></a>}}
\if{latex}{\out{\hypertarget{method-PersistentChat-reset_chat_history}{}}}
\subsection{Method \code{reset_chat_history()}}{
Reset the chat history
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persistent_chat-class$reset_chat_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersistentChat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PersistentChat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persistent_chat-class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
