% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Tidyprompt}
\alias{print.Tidyprompt}
\title{Print method for Tidyprompt (R6) objects}
\usage{
\method{print}{Tidyprompt}(x, ...)
}
\arguments{
\item{x}{A Tidyprompt R6 object}

\item{...}{Additional arguments (unused)}
}
\value{
The Tidyprompt object, invisibly.
}
\description{
This function prints a Tidyprompt object. It shows:
\itemize{
\item The base prompt
\item How many prompt wraps are applied
\item The fully constructed prompt text after modifications
}
}
