% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{prompt.prompt}
\alias{prompt.prompt}
\title{Method to create a prompt object from prompt}
\usage{
\method{prompt}{prompt}(input)
}
\arguments{
\item{input}{Input to create_prompt; the base prompt}
}
\value{
A prompt object
}
\description{
Method to create a prompt object from prompt
}
