% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{prompt.prompt}
\alias{prompt.prompt}
\title{Method to validate a prompt object}
\usage{
\method{prompt}{prompt}(input)
}
\arguments{
\item{input}{A prompt object}
}
\value{
A validated prompt object
}
\description{
Method to validate a prompt object
}
