% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{prompt_wrap.default}
\alias{prompt_wrap.default}
\title{Default method for prompt_wrap}
\usage{
\method{prompt_wrap}{default}(prompt, ...)
}
\arguments{
\item{prompt}{Input for the prompt object}

\item{...}{Additional arguments}
}
\value{
A prompt object with the new prompt wrap appended to it.
}
\description{
Attempts to create a prompt object from whatever is passed as 'prompt';
then calls the internal function to append the prompt wrap.
}
