% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{prompt_wrap}
\alias{prompt_wrap}
\title{Wrap a prompt with additional functionality}
\usage{
prompt_wrap(
  prompt,
  modify_fn = NULL,
  extraction_fn = NULL,
  validation_fn = NULL,
  max_retries = 10,
  type = c("unspecified", "mode", "tool")
)
}
\arguments{
\item{prompt}{A prompt object or a single string}

\item{modify_fn}{A function that takes the previous prompt text and returns the new prompt text}

\item{extraction_fn}{A function that takes the LLM response and attempts to extract a value from it.
Upon succesful extraction, the function should return the extracted value.
If the extraction fails, the function should return a feedback message which will be sent
back to the LLM (the feedback message should be of class 'llm_feedback').}

\item{validation_fn}{A function that takes the (extracted) LLM response and attempts to validate it.
Upon succesful validation, the function should return TRUE.
If the validation fails, the function should return a feedback message which will be sent
back to the LLM (the feedback message should be of class 'llm_feedback').#'}

\item{max_retries}{The maximum number of retries allowed for this prompt wrap;
each failure of extraction or validation will count as a retry.}

\item{type}{The type of prompt wrap; one of 'unspecified', 'mode', or 'tool'.
Types are used to determine the order in which prompt wraps are applied.
Typically, tools are applied first, then modes, then unspecified wraps.}
}
\value{
A prompt object with the new prompt wrap appended to it.
}
\description{
Wrap a prompt with additional functionality
}
