% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{prompt_wrap_internal}
\alias{prompt_wrap_internal}
\title{Internal function to append a prompt wrap to a prompt object}
\usage{
prompt_wrap_internal(
  prompt,
  modify_fn = NULL,
  extraction_fn = NULL,
  validation_fn = NULL,
  max_retries = 10,
  type = c("unspecified", "mode", "tool")
)
}
\arguments{
\item{prompt}{See prompt_wrap}

\item{modify_fn}{See prompt_wrap}

\item{extraction_fn}{See prompt_wrap}

\item{validation_fn}{See prompt_wrap}

\item{max_retries}{See prompt_wrap}

\item{type}{See prompt_wrap}
}
\value{
A prompt object with the new prompt wrap appended to it.
}
\description{
Internal function to append a prompt wrap to a prompt object
}
