% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_llm.R
\name{query_llm}
\alias{query_llm}
\title{...}
\usage{
query_llm(
  prompt,
  llm_provider = NULL,
  extractors = list(),
  validations = list(),
  max_retries = 10,
  verbose = TRUE
)
}
\arguments{
\item{prompt}{...}

\item{llm_provider}{...}

\item{extractors}{...}

\item{validations}{...}

\item{max_retries}{...}

\item{verbose}{....}
}
\value{
...
}
\description{
...
}
