% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_llm_provider.R
\name{request_llm_provider}
\alias{request_llm_provider}
\title{Make a request to an LLM provider}
\usage{
request_llm_provider(
  chat_history,
  request,
  stream = NULL,
  verbose = getOption("tidyprompt.verbose", TRUE),
  api_type = c("openai", "ollama")
)
}
\arguments{
\item{chat_history}{A data frame with 'role' and 'content' columns}

\item{request}{A 'httr2' request object with the URL, headers, and body}

\item{stream}{Logical indicating whether the API should stream responses.}

\item{verbose}{Logical indicating whether interactions should be printed to the console.}

\item{api_type}{API type, one of "openai" or "ollama".}
}
\value{
A list with the completed chat history and the HTTP request and response
objects
}
\description{
Make a request to an LLM provider
}
