% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_prompt.R
\name{send_prompt}
\alias{send_prompt}
\title{Send a prompt to a LLM provider}
\usage{
send_prompt(
  prompt,
  llm_provider,
  max_interactions = 10,
  verbose = getOption("tidyprompt.verbose", FALSE)
)
}
\arguments{
\item{prompt}{A prompt object or a single string}

\item{llm_provider}{'llm_provider' object}

\item{max_interactions}{Maximum number of interactions before stopping}

\item{verbose}{If the interaction with the LLM provider should be printed}
}
\value{
...
}
\description{
Send a prompt to a LLM provider
}
