% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_prompt.R
\name{send_prompt}
\alias{send_prompt}
\title{...}
\usage{
send_prompt(
  prompt,
  llm_provider = NULL,
  system_prompt = NULL,
  tool_functions = list(),
  extraction_functions = list(),
  validation_functions = list(),
  max_retries = 10,
  verbose = TRUE,
  extract_validate_mode = c("extraction_then_validation", "wrap_by_wrap")
)
}
\arguments{
\item{prompt}{...}

\item{llm_provider}{...}

\item{system_prompt}{...}

\item{tool_functions}{...}

\item{extraction_functions}{...}

\item{validation_functions}{...}

\item{max_retries}{...}

\item{verbose}{...}

\item{extract_validate_mode}{...}
}
\value{
...
}
\description{
...
}
