% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{set_llm_provider}
\alias{set_llm_provider}
\title{Set the LLM provider for a prompt list}
\usage{
set_llm_provider(prompt_wrap_or_list, llm_provider)
}
\arguments{
\item{prompt_wrap_or_list}{A single string, a prompt_wrap object, or a list
of prompt_wrap objects.}

\item{llm_provider}{llm_provider object to be used for evaluation
of the final prompt.}
}
\value{
A prompt list with the LLM provider set. The LLM provider
will be set for the base prompt in the list.
}
\description{
Set the LLM provider for a prompt list
}
