% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_system_prompt.R
\name{set_system_prompt}
\alias{set_system_prompt}
\title{Set system prompt}
\usage{
set_system_prompt(prompt, system_prompt)
}
\arguments{
\item{prompt}{A single string or a \code{\link[=tidyprompt]{tidyprompt()}} object}

\item{system_prompt}{A single character string representing the system prompt}
}
\value{
A \code{\link[=tidyprompt]{tidyprompt()}} with the system prompt set
}
\description{
Set the system prompt for a prompt. The system prompt will be added
as a message with role 'system' at the start of the chat history when
this prompt is evaluated by \code{\link[=send_prompt]{send_prompt()}}.
}
\details{
The system prompt will be stored in the \code{\link[=tidyprompt]{tidyprompt()}} object
as '$system_prompt'.
}
\seealso{
Other other_prompt_wraps: 
\code{\link{add_text}()}
}
\concept{other_prompt_wraps}
