% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_system_prompt.R
\name{set_system_prompt}
\alias{set_system_prompt}
\title{Set system prompt}
\usage{
set_system_prompt(prompt, system_prompt)
}
\arguments{
\item{prompt}{A prompt object}

\item{system_prompt}{A single character string representing the system prompt}
}
\value{
A prompt object with the system prompt set
}
\description{
Set the system prompt for a prompt
}
