% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt-package.R
\docType{package}
\name{tidyprompt-package}
\alias{tidyprompt-package}
\title{tidyprompt: Prompt Large Language Models ('LLMs') And Enhance Their Functionality}
\description{
‘tidyprompt’ is an R package to easily prompt large language models ('LLMs') and enhance their functionality. A framework is provided which allows users to construct LLM prompts using 'tidyverse'-inspired piping syntax. 'Prompt wraps' influence how a LLM handles the base prompt, through modifying the original text and applying extraction and validation functions to the LLM response. Users can choose from a library of pre-built prompt wraps or write their own. This allows for obtaining structured LLM output, with automatic feedback and retries if necessary. Additionally, the package supports LLM reasoning modes (e.g., chain of thought), autonomous R function calling, and R code generation and evaluation. 'tidyprompt' is compatible with any LLM provider that offers chat completions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjarkvandemerwe/tidyprompt}
  \item \url{https://tjarkvandemerwe.github.io/tidyprompt/}
  \item Report bugs at \url{https://github.com/tjarkvandemerwe/tidyprompt/issues}
}

}
\author{
\strong{Maintainer}: Luka Koning \email{l.koning@kennispunttwente.nl} [copyright holder]

Authors:
\itemize{
  \item Tjark Van de Merwe \email{t.vandemerwe@kennispunttwente.nl} [copyright holder]
}

Other contributors:
\itemize{
  \item Kennispunt Twente \email{info@kennispunttwente.nl} [funder]
}

}
\keyword{internal}
