% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt-package.R
\docType{package}
\name{tidyprompt-package}
\alias{tidyprompt-package}
\title{tidyprompt: Prompt Large Language Models and Enhance Their Functionality}
\description{
Easily construct prompts and associated logic for interacting with large language models ('LLMs'). 'tidyprompt' introduces the concept of prompt wraps, which are building blocks that you can use to quickly turn a simple prompt into a complex one. Prompt wraps do not just modify the prompt text, but also add extraction and validation functions that will be applied to the response of the LLM. This ensures that the user gets the desired output. 'tidyprompt' can add various features to prompts and their evaluation by LLMs, such as structured output, automatic feedback, retries, reasoning modes, autonomous R function calling, and R code generation and evaluation. It is designed to be compatible with any LLM provider that offers chat completion.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjarkvandemerwe/tidyprompt}
  \item \url{https://tjarkvandemerwe.github.io/tidyprompt/}
  \item Report bugs at \url{https://github.com/tjarkvandemerwe/tidyprompt/issues}
}

}
\author{
\strong{Maintainer}: Luka Koning \email{l.koning@kennispunttwente.nl} [copyright holder]

Authors:
\itemize{
  \item Tjark Van de Merwe \email{t.vandemerwe@kennispunttwente.nl} [copyright holder]
}

Other contributors:
\itemize{
  \item Kennispunt Twente \email{info@kennispunttwente.nl} [funder]
}

}
\keyword{internal}
