% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt-package.R
\docType{package}
\name{tidyprompt-package}
\alias{tidyprompt-package}
\title{tidyprompt: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjarkvandemerwe/tidyprompt}
  \item Report bugs at \url{https://github.com/tjarkvandemerwe/tidyprompt/issues}
}

}
\author{
\strong{Maintainer}: Luka Koning \email{l.koning@kennispunttwente.nl}

Authors:
\itemize{
  \item Tjark Van de Merwe \email{t.vandemerwe@kennispunttwente.nl}
}

}
\keyword{internal}
