% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{tidyprompt}
\alias{tidyprompt}
\title{Create a Tidyprompt object}
\usage{
tidyprompt(input)
}
\arguments{
\item{input}{Input to prompt. If a character string is passed,
a new prompt object will be created with that character string as the base prompt.}
}
\value{
A Tidyprompt object
}
\description{
Create a Tidyprompt object
}
\seealso{
\code{\link[=prompt_wrap]{prompt_wrap()}}

Other tidyprompt: 
\code{\link{construct_prompt_text}()},
\code{\link{get_prompt_wraps}()},
\code{\link{is_tidyprompt}()},
\code{\link{tidyprompt-class}}
}
\concept{tidyprompt}
