% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_wrap.R
\name{tidyprompt}
\alias{tidyprompt}
\title{Create a base prompt wrap object}
\usage{
tidyprompt(prompt_text, ...)
}
\arguments{
\item{prompt_text}{Text of the prompt}

\item{...}{Additional arguments to be passed to create_prompt_wrap}
}
\value{
A prompt wrap object
}
\description{
Function to create a prompt wrap object which will be at the base of the
prompt list. Typically, prompt wrappers will be added to this base prompt.
}
\examples{
base_prompt <- tidyprompt("Enter a number between 1 and 10")

# Note that if your base_prompt is only text, you do not need to use tidyprompt.
# You can also use the other functions directly with a string of text as the
# base prompt.
}
