% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{tidyprompt.tidyprompt}
\alias{tidyprompt.tidyprompt}
\title{Validate and return tidyprompt}
\usage{
\method{tidyprompt}{tidyprompt}(input)
}
\arguments{
\item{input}{A tidyprompt object}
}
\value{
A validated tidyprompt object
}
\description{
Validate and return tidyprompt
}
