% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprompt.R
\name{validate_tidyprompt}
\alias{validate_tidyprompt}
\title{Validate tidyprompt}
\usage{
validate_tidyprompt(tidyprompt)
}
\arguments{
\item{tidyprompt}{A tidyprompt object}
}
\value{
TRUE if the tidyprompt is valid, otherwise an error is thrown
}
\description{
Validate tidyprompt
}
